Implementation of north angle calculations (tlit, drift compensation, low-pass filtration) MVP - Model layer added EventBus - communication between DataProvider and CompassPresenter RxJava - long calculations handling

+++ app/src/main/java/slosar/example/compassapp/CompassDisplay/ICompassView.java
@@ -4,7 +4,7 @@ package slosar.example.compassapp.CompassDisplay;
  * Created by Rafal on 2016-04-12.
  */
 interface ICompassView {
-    void setCompassAngle(double currentAngle, double newAngle);
+    void setCompassAngle(float currentAngle, float newAngle);
 
-    void setDirectionAngle(double currentAngle, double newAngle);
+    void setDirectionAngle(float currentAngle, float newAngle);
 }
